Clean up characters routes file by creating charactersController
